refactor(schedule): move schedule filling out of views
fill_next_weeks_agenda and the jukebox filler were never part of the
request path -- both are invoked only by management commands, and thus
by the nightly CronJobs. Living in views.py made them look like view
code and gave the module a logger and a WeeklySlot import that no view
used.

Moved verbatim to agenda/scheduling.py: no logic changes, and the
callers (two management commands, three test modules) now import from
there. views.py keeps only view code.
